refactor remarkFrontmatterToProps
