+++ packages/tools/docs-tools/src/remarkFrontmatterToProps/utils/getFileName.ts
@@ -1,0 +1,6 @@
+import type { IFile } from './../../types';
+
+export const getFileName = (file: IFile): string => {
+  if (file.history.length === 0) return '';
+  return file.history[0];
+};
