+++ packages/tools/docs-tools/vitest.config.ts
@@ -12,10 +12,10 @@ export default mergeConfig(
       coverage: {
         provider: 'v8',
         thresholds: {
-          lines: 0,
-          functions: 0,
-          branches: 0,
-          statements: 0,
+          lines: 26.02,
+          functions: 60,
+          branches: 74.68,
+          statements: 26.02,
         },
       },
     },
